Issue #54 - Reputation permissions for reviews and responses are in place.
Still need to do permissions on publishing.  But we've officially hit
the point where it's time to optimize.  The site becoming painfully
slow, managing all those fields with reputation attached for each user
is very quickly challenging the browser.  So, we need to optimize it!

+++ server/controllers/votes.js
@@ -5,6 +5,9 @@
  *
  ******************************************************************************/
 const ReputationService = require('../services/reputation')
+const ReputationPermissionService = require('../services/ReputationPermissionService')
+
+const ControllerError = require('../errors/ControllerError')
 
 /**
  *
@@ -15,6 +18,7 @@ module.exports = class VoteController {
         this.database = database
         this.logger = logger
         this.reputationService = new ReputationService(database, logger)
+        this.reputationPermissionService = new ReputationPermissionService(database, logger)
     }
 
 
@@ -24,17 +28,11 @@ module.exports = class VoteController {
      * Return a JSON array of all votes for this paper in the database.
      */
     async getVotes(request, response) {
-        try {
-            const paper_id = request.params.paper_id
-            const results = this.database.query('SELECT paper_id as "paperId", user_id as "userId", score from paper_votes where paper_id=$1', [ paper_id ])
+        const paper_id = request.params.paper_id
+        const results = await this.database.query('SELECT paper_id as "paperId", user_id as "userId", score from paper_votes where paper_id=$1', [ paper_id ])
 
-            const votes = results.rows
-            return response.status(200).json(votes)
-        } catch (error) {
-            this.logger.error(error)
-            response.status(500).json({ error: 'unknown' })
-            return
-        }
+        const votes = results.rows
+        return response.status(200).json(votes)
     }
 
     /**
@@ -43,31 +41,44 @@ module.exports = class VoteController {
      * Create a new vote in the database from the provided JSON.
      */
     async postVotes(request, response) {
+        // Only an authenticated user can vote.
+        if ( ! request.session.user ) {
+            throw new ControllerError(403, 'not-authorized', `An attempt to vote by an un-authenticated user.`)
+        }
+
+        // Ensure the user has enough reputation to be able to vote on this
+        // paper.
+        const canVote = this.reputationPermissionService.canReferee(request.session.user.id, request.params.paper_id)
+        if ( ! canVote ) {
+            throw new ControllerError(403, 'not-enough-reputation', `User(${request.session.user.id}) attempted to vote on paper(${request.params.paper_id}) with out enough reputation.`)
+        }
+
         const vote = request.body
 
-        try {
-            const results = await this.database.query(`
+        // Ensure that the user is casting their own vote and not someone
+        // else's.
+        if ( vote.userId != request.session.user.id ) {
+            throw new ControllerError(400, 'user-mismatch', `User(${request.session.user.id}) attempted to post a vote for different User(${vote.userId}).`)
+        }
+
+        const results = await this.database.query(`
                 INSERT INTO paper_votes (paper_id, user_id, score, created_date, updated_date) 
                     VALUES ($1, $2, $3, now(), now()) 
                 RETURNING paper_id as "paperId", user_id as "userId", score
                 `, 
-                [ request.params.paper_id, vote.userId, vote.score]
-            )
-            if ( results.rows.length == 0 ) {
-                this.logger.error('Failed to insert a vote.')
-                return response.status(500).json({error: 'unknown'})
-            }
+            [ request.params.paper_id, vote.userId, vote.score]
+        )
 
-
-            const returnVote = results.rows[0]
-
-            await this.reputationService.incrementReputationForPaper(returnVote.paperId, returnVote.score)
-
-            return response.status(201).json(returnVote)
-        } catch (error) {
-            this.logger.error(error)
-            return response.status(500).json({error: 'unknown'})
+        if ( results.rows.length == 0 ) {
+            throw new ControllerError(500, 'server-error', `Failed to insert a vote for user(${vote.userId}) and paper(${request.params.paper_id}).`)
         }
+
+
+        const returnVote = results.rows[0]
+
+        await this.reputationService.incrementReputationForPaper(returnVote.paperId, returnVote.score)
+
+        return response.status(201).json(returnVote)
     }
 
     /**
@@ -76,21 +87,16 @@ module.exports = class VoteController {
      * Get details for a single vote in the database.
      */
     async getVote(request, response) {
-        try {
-            const paper_id = request.params.paper_id
-            const user_id = request.params.user_id
+        const paper_id = request.params.paper_id
+        const user_id = request.params.user_id
 
-            const results = this.database.query('SELECT paper_id as "paperId", user_id as "userId", score from paper_votes where paper_id=$1 and user_id=$2', [ paper_id, user_id ])
+        const results = await this.database.query('SELECT paper_id as "paperId", user_id as "userId", score from paper_votes where paper_id=$1 and user_id=$2', [ paper_id, user_id ])
 
-            if ( ! vote ) {
-                return response.status(404).json({})
-            }
-
-            return response.status(200).json(vote)
-        } catch (error) {
-            this.logger.error(error)
-            return response.status(500).json({ error: 'unknown' })
+        if ( ! vote ) {
+            return response.status(404).json({})
         }
+
+        return response.status(200).json(vote)
     }
 
     /**
@@ -99,28 +105,39 @@ module.exports = class VoteController {
      * Replace an existing vote wholesale with the provided JSON.
      */
     async putVote(request, response) {
-        try {
-            const vote = request.body
+        // Only an authenticated user can change a vote.
+        if ( ! request.session.user ) {
+            throw new ControllerError(403, 'not-authorized', `An attempt to vote by an un-authenticated user.`)
+        }
 
-            // Update the vote.
-            const results = await this.database.query(`
+        // Check to ensure the user has enough reputation to change this vote.
+        const canVote = this.reputationPermissionService.canReferee(request.session.user.id, request.params.paper_id)
+        if ( ! canVote ) {
+            throw new ControllerError(403, 'not-enough-reputation', `User(${request.session.user.id}) attempted to vote on paper(${request.params.paper_id}) with out enough reputation.`)
+        }
+
+        // Ensure that the user changing the vote is the same one logged in.
+        if ( request.params.user_id != request.session.user.id) {
+            throw new ControllerError(400, 'user-mismatch', `User(${request.session.user.id}) attempted to post a vote for different User(${vote.userId}).`)
+        }
+
+        const vote = request.body
+
+        // Update the vote.
+        const results = await this.database.query(`
                 UPDATE paper_votes 
                     SET score = $1, updated_date = now() 
                 WHERE paper_id = $2 and user_id=$3
                 RETURNING paper_id as "paperId", user_id as "userId", score
                 `,
-                [ vote.score, request.params.paper_id, request.params.user_id ]
-            )
+            [ vote.score, request.params.paper_id, request.params.user_id ]
+        )
 
-            if (results.rowCount == 0 && results.rows.length == 0) {
-                return response.status(404).json({error: 'no-resource'})
-            }
-
-            response.status(200).json(results.rows)
-        } catch (error) {
-            this.logger.error(error)
-            response.status(500).json({error: 'unknown'})
+        if (results.rowCount == 0 && results.rows.length == 0) {
+            return response.status(404).json({error: 'no-resource'})
         }
+
+        response.status(200).json(results.rows)
     }
 
 
@@ -130,20 +147,31 @@ module.exports = class VoteController {
      * Delete an existing vote.
      */
     async deleteVote(request, response) {
-        try {
-            const results = await this.database.query(
-                'delete from paper_votes where paper_id = $1 and user_id = $2',
-                [ request.params.paper_id, request.params.user_id ]
-            )
-
-            if ( results.rowCount == 0) {
-                return response.status(404).json({error: 'no-resource'})
-            }
-
-            return response.status(200).json(null)
-        } catch (error) {
-            this.logger.error(error)
-            return response.status(500).json({error: 'unknown'})
+        // Only an authenticated user can change a vote.
+        if ( ! request.session.user ) {
+            throw new ControllerError(403, 'not-authorized', `An attempt to vote by an un-authenticated user.`)
         }
+
+        // Check to ensure the user has enough reputation to change this vote.
+        const canVote = this.reputationPermissionService.canReferee(request.session.user.id, request.params.paper_id)
+        if ( ! canVote ) {
+            throw new ControllerError(403, 'not-enough-reputation', `User(${request.session.user.id}) attempted to vote on paper(${request.params.paper_id}) with out enough reputation.`)
+        }
+
+        // Ensure that the user changing the vote is the same one logged in.
+        if ( request.params.user_id != request.session.user.id) {
+            throw new ControllerError(400, 'user-mismatch', `User(${request.session.user.id}) attempted to post a vote for different User(${vote.userId}).`)
+        }
+
+        const results = await this.database.query(
+            'delete from paper_votes where paper_id = $1 and user_id = $2',
+            [ request.params.paper_id, request.params.user_id ]
+        )
+
+        if ( results.rowCount == 0) {
+            return response.status(404).json({error: 'no-resource'})
+        }
+
+        return response.status(200).json(null)
     }
 }
